+++ src/context/HabitsContext.tsx
@@ -29,7 +29,9 @@ interface HabitsContextType {
   refreshVisibleHabits: () => Promise<void>;
 }
 
-const HabitsContext = createContext<HabitsContextType | undefined>(undefined);
+export const HabitsContext = createContext<HabitsContextType | undefined>(
+  undefined,
+);
 
 export const HabitsProvider = ({ children }: { children: ReactNode }) => {
   const [databaseHabits, setDatabaseHabits] = useState<Habit[]>([]);
@@ -141,12 +143,4 @@ export const HabitsProvider = ({ children }: { children: ReactNode }) => {
       {children}
     </HabitsContext.Provider>
   );
-};
-
-export const useHabits = () => {
-  const context = useContext(HabitsContext);
-  if (context === undefined) {
-    throw new Error("useHabits must be used within a HabitsProvider");
-  }
-  return context;
 };
