+++ src/context/useHabits.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { HabitsContext } from "./HabitsContext";
+
+export const useHabits = () => {
+  const context = useContext(HabitsContext);
+  if (context === undefined) {
+    throw new Error("useHabits must be used within a HabitsProvider");
+  }
+  return context;
+};
